backend/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-api-python-client==2.93.0
pydantic==1.10.11
python-dotenv==1.0.0 